OXDEV-846 Fix collation errors during setup
The migration wrapper should use the same database connection parameters like the shop. This solution with hardcoding the values
was chosen as there is no global class for constants right now. In the future, this can be easily refactored.
Probably those parameters should not be configurable via config.inc.php as the shop won't work as expected setting anything else than the defaults.